samples: Bluetooth: Rename BLE to Bluetooth (LE) where applicable
The BLE acronym is not an official description of Bluetooth
LE, and the Bluetooth SIG only ever refers to it as Bluetooth
Low Energy or Bluetooth LE, so Zephyr should as well.

This commit does not change any board or vendor specific
samples, and the term BLE may still be used in those.
It will be up to the vendors to update it if they want,
since many of them are using the term BLE in their
products.

(cherry picked from commit 3b6a15a24d377ee16517696464642e637e671c9a)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 3b6a15a24d377ee16517696464642e637e671c9a
Cr-Build-Id: 8723142769276654049
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8723142769276654049
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Id6e36c199b03515fcd936bdec8262fe7a16eea69
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6257420
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>