net: Update IP address refcount properly when address already exists
If an IP address already exists when it is tried to be added to the
network interface, then just return it but update ref count if it was
not updated. This could happen if the address was added and then removed,
but for example an active connection was still using it and keeping the
ref count > 0. In this case we must update the ref count so that the IP
address is not removed if the connection is closed.

Fixes #85380

(cherry picked from commit ae052217625881a581e7585f3603e1dbbc298571)

Original-Signed-off-by: Jukka Rissanen <jukka.rissanen@nordicsemi.no>
GitOrigin-RevId: ae052217625881a581e7585f3603e1dbbc298571
Cr-Build-Id: 8723142769276654049
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8723142769276654049
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I5db6a302476913f11fc545f1deee38f9096a40d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6257405
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>