Bluetooth: Host: Make bt_le_addr_is_bonded public
Some GATT services and profiles define specific behavior
based on whether the remote device is bonded or not.
The internal function, bt_addr_le_is_bonded, is the
only function to do this, but it was kept internal,
and could thus not be used for those services
without including hci_core.h.

The function has been moved to the public API
so that application can determine if a remote
address is bonded or not, and has been renamed
to not use the bt_addr namespace, but rather the
bt_le.

(cherry picked from commit 59184272784b8ad26efbfdcfeed9cbf921599108)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: 59184272784b8ad26efbfdcfeed9cbf921599108
Cr-Build-Id: 8720515456456009505
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8720515456456009505
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I34c3384680d75595ec3c4420a467a8887b94e8aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6351625
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>