drivers: mm: rat: Fix virt and phys NULL checks
Do not check virt being NULL as 0 is a valid address that can be
translated. Phys on the other hand is an output pointer that will be
assigned to and so cannot be NULL, check this instead to prevent a
NULL pointer dereference later.

Note that phys is allowed to be NULL, so this is not an invalid argument.
It can set to NULL when the caller wants to check if a virtual address is
mapped but does not care about the address itself. In that case -EFAULT
should be returned if the virtual address is not mapped, 0 otherwise.
For RAT, all addresses not translated are still valid and mapped as
pass-through, so we always return 0 in this check case.

(cherry picked from commit 7949ffd7d8f0c766f8285ccde38e337a7f5ef40f)

Original-Signed-off-by: Andrew Davis <afd@ti.com>
GitOrigin-RevId: 7949ffd7d8f0c766f8285ccde38e337a7f5ef40f
Cr-Build-Id: 8720606055414277025
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8720606055414277025
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: If0b5b5d44c3765af776f2bdc98382f4b12ec7bde
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6346832
Tested-by: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>