boards nrf54l15bsim: Do not work around peripheral clock issue
In 923d313a04726f6475b19ef803eb8b5a080e91d7 the clock frequency in DTS
for the UART00 was fixed, but not for the simulated target. This was
likely due to the HW models modeling it as 16MHz instead of 128MHz for
this particular one as it is in reality.

Now that the HW models have been fixed, let's let this clock be
configured like for real HW.

(cherry picked from commit 5e7df92082ccb084ba13c2abf98f8512a00ad192)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: 5e7df92082ccb084ba13c2abf98f8512a00ad192
Cr-Build-Id: 8722916276808742417
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8722916276808742417
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ib013c176f10f0a8c3485e8f86a1425b7d779c8b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6272062
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>