drivers: i3c: cleanup target config descrition and add inline helpers
It shouldn't be possible to just 'enable' target mode of a device.
It is required by the specification to perform a handoff or request to
become a target or controller from the active controller. Not to just flip
a switch internally. Can the parameter from `enable` to `enabled` to report
if it is currently a target or a controller otherwise.

Also, add inline helpers for `i3c_config_get` and `i3c_configure` to ensure
the proper struct is passed in.

(cherry picked from commit b495907ae58bb333e3b1cc8ba526a42277522b7b)

Original-Signed-off-by: Ryan McClelland <ryanmcclelland@meta.com>
GitOrigin-RevId: b495907ae58bb333e3b1cc8ba526a42277522b7b
Cr-Build-Id: 8721013740965942577
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8721013740965942577
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I504bd78a5197ee84c880610273d590a02270dea5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6337236
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>