drivers: pinctrl: silabs: Fix pin deallocation from digital bus
Fix the scenario where a pinctrl node intends to deallocate a pin
from a peripheral. If the GPIO mode is disabled the DBUS route
should be cleared, not set. This allows reuse of a pin for other
purposes when a driver is suspended and the pinctrl sleep state is
applied, as GPIOs are typically disabled in the sleep state.

(cherry picked from commit 23c5144f9ca08cebbb9ae090566c5ad724e1baf9)

Original-Signed-off-by: Aksel Skauge Mellbye <aksel.mellbye@silabs.com>
GitOrigin-RevId: 23c5144f9ca08cebbb9ae090566c5ad724e1baf9
Cr-Build-Id: 8722553887981252641
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8722553887981252641
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I6b0ebe863b5960a9212233a9f5822473ef632147
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6279995
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>