native FUSE FS access: Support any libC and fix random crashes
Split the fuse FS driver into 2 parts: A top built in the embedded side,
with the embedded libC, and a bottom built in the runner side with the
host libC.
The error returns are converted to match the host libC.

Also, before the host FUSE thread, which is asynchronous to Zephyr was
calling directly into the Zephyr filesystem code, which resulted quite
often if catastrophic failures or corruption of the Zephyr state.
This is now fixed by having the FUSE thread queue requests to a Zephyr
thread, which will be handled in the embedded side in a coherent way.
This adds a slightly noticeable overhead, but the performance is still
acceptable.

(cherry picked from commit ac24d2277e021a92d435efc45f9121e00970d0cd)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: ac24d2277e021a92d435efc45f9121e00970d0cd
Cr-Build-Id: 8720945792751178849
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8720945792751178849
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I984479f9e65f9acddc90a1c66227c698f4e7d418
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6337738
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>