drivers: dma: dma_xilinx_axi_dma: Fix up IRQ locking
The way the driver was storing IRQ numbers for later use in the various
locking modes was not correct, causing the wrong IRQs to potentially be
disabled/enabled in some modes. Refactor the way this is done to be
cleaner, and also the way the different locking modes are implemented in
order to ensure that all modes receive compile test coverage.

Also, ensure that the IRQ for the RX or TX channel is always disabled
during the execution of the corresponding ISR, to prevent it from being
preempted by itself if the DMA core raises another interrupt during the
ISR execution.

(cherry picked from commit 0345fc9d2a47c408feea8f23fa1905bff42335e1)

Original-Signed-off-by: Robert Hancock <robert.hancock@calian.com>
GitOrigin-RevId: 0345fc9d2a47c408feea8f23fa1905bff42335e1
Cr-Build-Id: 8692226556012859633
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8692226556012859633
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: If50414679ddce56df833616c34be202942d1b9cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/7493837
Tested-by: Dawid Niedźwiecki <dawidn@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>
Commit-Queue: Dawid Niedźwiecki <dawidn@google.com>
Reviewed-by: Dawid Niedźwiecki <dawidn@google.com>