soc: nordic: nrf54h: Add option to use NFCT pins as GPIOs
Protection circuit must be disabled to use NFCT antenna pins
as GPIOs. It can be done by adding nfct-pins-as-gpios to nfct
node in the devicetree in cpuapp. Node must be disabled as
NFCT is not used. In legacy platforms same property was added
to uicr node because that information was stored in UICR. In
nrf54h20 it is not part of UICR so property is part of nfct
node.

(cherry picked from commit 7249fac80d60b84d2ba9c2a1de6d42d8b3d9bd0c)

Original-Signed-off-by: Krzysztof Chruściński <krzysztof.chruscinski@nordicsemi.no>
GitOrigin-RevId: 7249fac80d60b84d2ba9c2a1de6d42d8b3d9bd0c
Cr-Build-Id: 8720583405886367041
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8720583405886367041
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I2f0f89a1c15b9787019f8d376a674adacb7823d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6348427
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>