Bluetooth: userchan: Support other libCs
Refactor the userchan driver into a top and a bottom part.
The bottom is the one which interacts with the host and is built
with the host libC, while the top is built with the embedded code
and whatever libC that is built with.

Errors (errno) is converted between the top and bottom to ensure they
are coherent with the local libC.

(cherry picked from commit b2b6f3c314382cb33deac576f0ff9af565e37004)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: b2b6f3c314382cb33deac576f0ff9af565e37004
Cr-Build-Id: 8720787248423163057
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8720787248423163057
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I1a369946872f09aa15a66807704a76243cb1c81b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6338671
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>