drivers: sensor: ti: opt3001: add trigger support
The ti,opt3001 ambient light sensor is capable of asserting its interrupt
gpio upon conversion completion. The interrupt gpio can then be used to
trigger a reading of the conversion result automatically.
Due to this, it makes sense for the driver to implement the sensor trigger
functionality.

Add a devicetree property, int-gpios, to the ti,opt3001 bindings to allow
describing the gpio used for getting interrupts from the ti,opt3001.

Implement the sensor trigger functionality with options for using the
global worker thread or letting the ti,opt3001 create its own thread for
servicing the interrupts.
In both cases, when an interrupt is received the driver will read a new
measurement from the chip and clear the interrupt. The sample is then ready
to be read by the user of the sensor.

When manually triggering a sample_fetch, the trigger handler is not
invoked.

(cherry picked from commit 1a3dd63dde40edaeb1bbea86913913ac35341d2a)

Original-Signed-off-by: Emil Dahl Juhl <emdj@bang-olufsen.dk>
GitOrigin-RevId: 1a3dd63dde40edaeb1bbea86913913ac35341d2a
Cr-Build-Id: 8718227884099254305
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8718227884099254305
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I2d19a0e4812436adcbd93787eac54f72a1a5972a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6439113
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>