Bluetooth: SDP: Ignore unsupported UUID
In current implementation, if the UUID of request is unknown, the
local variable `u` will not be updated, and the unknown UUID `u`
is used as the target to search it in the SDP record list. It may
result in irrelevant record being replied to the requester.

Ignore the unsupported UUID to fix the issue.

(cherry picked from commit 86b765c0a940370af88e4fa31ac9778d9fb0986b)

Original-Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>
GitOrigin-RevId: 86b765c0a940370af88e4fa31ac9778d9fb0986b
Cr-Build-Id: 8720991092936174337
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8720991092936174337
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Iee476b0b2fbed49acc711ec2eb353b76f70ecfa8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6337649
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>