boards: renesas: Modify the pinsel for adc0 on ek_ra4l1
Currently, we set P002 (AN000) as the default pin selection
for ADC0 on ek_ra4l1. However, after testing the ADC function
with this pin selection, the ADC function does not work
on this pin.

The main reason is P002 is used for the I2C I/O Port Expander
and is connected to a 10kΩ pull-up resistor.

For the solution, we will select the P003 (AN001)
as default pinsel for adc0

(cherry picked from commit 1aa48e624752c53cbfca03754dcdaccd340bab81)

Original-Signed-off-by: Khoa Nguyen <khoa.nguyen.xh@renesas.com>
GitOrigin-RevId: 1aa48e624752c53cbfca03754dcdaccd340bab81
Cr-Build-Id: 8720606055414277025
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8720606055414277025
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ib8b5d50abbd98aec61fe8f278f24ace2371700bc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6346843
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>