llext: support alignment requirements for sections
This patch adds support for alignment requirements for sections in the
loader. The alignment requirements are taken from the ELF file and
checked when the section is mapped to a memory region and when the
region is copied to memory (or directly used from the ELF buffer).

Supporting a larger alignment requirement than the region's current
alignment is done by adjusting the region's start address. This needs
care to avoid detecting false overlaps with other regions and to ensure
that the region address and size reported by the inspection APIs are
correct. The extra offset needed to re-adjust these values is stored in
the 'sh_info' field of the region descriptor.

(cherry picked from commit 959f1b352d03c007e85d0bfbb8c2371c570a7080)

Original-Signed-off-by: Luca Burelli <l.burelli@arduino.cc>
GitOrigin-RevId: 959f1b352d03c007e85d0bfbb8c2371c570a7080
Cr-Build-Id: 8720107771778692097
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8720107771778692097
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ic1289f97b55340c7c464e31219e0c168c5499c13
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6361846
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>