Bluetooth: SSP: Reply a negative rsp if binding flags are mismatched
If the remote is in bondable mode, but the local is in non-bondable
mode, the local host shall respond to an IO capability request with
a negative response.

In current implementation, it does not check the bonding flags of the
both sides are consistent.

Fix the issue by checking the consistency of bonding flags of the both
sides. If they are not consistent, send a negative IO capability
response with the reason pairing not allowed.

(cherry picked from commit dcd8bcb88d2dd246ef6716819952d6ccb0cfd4de)

Original-Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>
GitOrigin-RevId: dcd8bcb88d2dd246ef6716819952d6ccb0cfd4de
Cr-Build-Id: 8720764599659500817
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8720764599659500817
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I5a7014eed9d6a3809fbf5e99da4f318f0dd15e74
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6340174
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>