boards: nrf: Default to nRF Util for flashing
Nordic has a new tool, nRF Util (`nrfutil`) which is replacing the
previous set of command-line tools (particularly `nrfjprog`) that Zephyr
had been using by default in many of the boards to flash the firmware
onto the device.

Since nRF Util is already mature and is going to be replacing the existing
tools, including `nrfjprog`, it is now a good time to change the default
flash tool to reflect this fact.

Note that `nrfjprog` support is not being removed at all, but this may
be a breaking change for users that do not have nRF Util installed. For
this reason we also add a note in the migration guide.

More info:
https://devzone.nordicsemi.com/nordic/nordic-blog/b/blog/posts/nrf-util-unified-command-line-utility

Installation:
https://www.nordicsemi.com/Products/Development-tools/nrf-util

Documentation:
https://docs.nordicsemi.com/bundle/nrfutil/page/README.html

(cherry picked from commit 9fc84e78e73d80935581bbbdd16885b2b4996718)

Original-Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>
GitOrigin-RevId: 9fc84e78e73d80935581bbbdd16885b2b4996718
Cr-Build-Id: 8720741950543301233
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8720741950543301233
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ice6d06d8cbc4bc4edbebb3bac36de884cdee773c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6341355
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>