Bluetooth: SSP: Fix MITM flag incorrect issue in pairing
There is a case that if the local is a peripheral and a L2CAP server
with the security level 3，the mitm should be set to make sure the
security level can reach level 3.

But in current implementation, the MITM is not set because the MITM
of `Authentication_Requirements` parameter of remote is not set. And
the security level will be level 2 after the security pairing
procedure complete.

Add a function `bt_l2cap_br_get_max_sec_level` to
get the max level of the registered servers.
`
Force set the MITM bit if the max level is bigger than level 2 and
pairing method is not `just works`.

(cherry picked from commit f349976937012d1d6d61e12e8ba9ada3ca812daa)

Original-Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>
GitOrigin-RevId: f349976937012d1d6d61e12e8ba9ada3ca812daa
Cr-Build-Id: 8720606055414277025
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8720606055414277025
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I940b2d1935e1c86ef95f8ee4aaf6a00b55ddc8d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6346845
Tested-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>