Bluetooth: CSIP: Add support for dynamically setting set size
The set size can now be dynamically set and notified.
The rank is added as a argument in the case that changing
the set size, also affects the device's rank, as ranks
in a coordinated set needs to be continuous.

The set coordinator implementation has been updated
to support receiving the new set size, and providing
this information to the upper layers.

This commit adds a babblesim test for the new API,
as well as a shell command.

(cherry picked from commit d19abff47665165243c78aa8801ab302968487fa)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: d19abff47665165243c78aa8801ab302968487fa
Cr-Build-Id: 8719858630320399921
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8719858630320399921
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Ib8c43403b686eeb374adc4eec7283ffc2d47d04b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6377394
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>