drivers: udc_dwc2: Reset core on shutdown
OUT endpoint 0 cannot be disabled and therefore the only way to forcibly
reclaim the buffer is to reset the core. The reset does not finish if
PHY clock is not running, but just triggering the reset seems to be
enough to be able to reclaim the buffer.

(cherry picked from commit 2395452656099afee86d642c063b582df6b0d6c9)

Original-Signed-off-by: Tomasz Moń <tomasz.mon@nordicsemi.no>
GitOrigin-RevId: 2395452656099afee86d642c063b582df6b0d6c9
Cr-Build-Id: 8719496240540423633
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8719496240540423633
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I459ae803ab009a01d177e6060bc7e70c02c49e06
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6388500
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>