settings: zms: load only subtree in the argument
If the subtree argument is not NULL in the settings_load function, load
only the setting entry that corresponds to that subtree.
If the subtree argument is NULL or it is not found in the storage, load
all the existing entries in the storage.

(cherry picked from commit ae461510a7660c8da6541790925ce0af630bba21)

Original-Signed-off-by: Riadh Ghaddab <rghaddab@baylibre.com>
GitOrigin-RevId: ae461510a7660c8da6541790925ce0af630bba21
Cr-Build-Id: 8714785199454382401
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8714785199454382401
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I61625b53356ac90da6ed672c529a5bd3162deef7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6553789
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>