xtensa: set pointer to privileged stack only if user mode
When initializing the stack at thread creation, we should not
set the pointer to privileged stack pointer yet as the thread
can be a kernel thread. Only when a thread is transitioning to
user mode, then we need to set the pointer to point to
the privileged stack. This is a purely semantic change and
should not affect any functionality.

(cherry picked from commit dcf75319e2accaa6e075d4e4566336e261548979)

Original-Signed-off-by: Daniel Leung <daniel.leung@intel.com>
GitOrigin-RevId: dcf75319e2accaa6e075d4e4566336e261548979
Cr-Build-Id: 8717389862523825937
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8717389862523825937
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I679db0ceccd3a6cd85061e2a86cd191292c3dcb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6466399
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>