drivers: i3c: i3c_dw: dynamically attach I2C devices on transfer
The maximum number of attached devices is limited by the maximum
number of entries in the device address table. For I3C devices
these have to by allocated permanently when devices are attached
but for I2C devices we can dynamically allocate an entry in the
table on a I2C transfer and free it after the transfer is completed.
This allows the maximum number of I2C + I3C devices on a bus to be
larger than the address table size as long as the number of I3C
devices is maxdevs - 1.

(cherry picked from commit b818a7662d17c9203290736a73c61e214593018d)

Original-Signed-off-by: Corey Wharton <xodus7@cwharton.com>
GitOrigin-RevId: b818a7662d17c9203290736a73c61e214593018d
Cr-Build-Id: 8720447510340548033
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8720447510340548033
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I4b36e0f3936ce3ea75875c645e049deb14e0e725
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6356587
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>