drivers: udc_dwc2: Disable endpoint while hibernated
When the endpoint is disabled while the core is hibernated, there are
timeouts waiting for interrupts. It is not clear how the stack should
behave when class and/or application wants to disable the endpoint while
device is suspended. The problem was originally observed when the
endpoints were disabled as a result of usbd_disable() call.

Avoid the timeouts by modifying the backup values instead of the real
registers (which are not accessible when hibernated).

Split the 32-bit txf_set variable into two 16-bit variables (txf_set and
pending_tx_flush) because maximum number of TxFIFO instances is 16.
The txf_set variable is used as-is, while the pending_tx_flush is used
to keep track of TxFIFOs that have to be flushed on hibernation exit.

(cherry picked from commit 2be960ad2b33ce1e5af50a2337de80600ab5eea5)

Original-Signed-off-by: Tomasz Moń <tomasz.mon@nordicsemi.no>
GitOrigin-RevId: 2be960ad2b33ce1e5af50a2337de80600ab5eea5
Cr-Build-Id: 8719496240540423633
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8719496240540423633
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I377fad582cfe96f74b185325919e20d75755d4c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6388498
Bot-Commit: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Commit-Queue: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>