drivers: i3c: stm32: fix clock init for i2c fast mode plus
The logic of clock initialization for i2c fast mode (FM) and fast
mode plus (FMP) is as follows:
 1 compute how many system clock cycles for SCL to be low
 2 compute how many system clock cycles for SCL to be high by
   subtracting the low duration computed above from the SCL period
 3 verify the high duration computed in 2 is larger than a minimum
The bug is that the step 3 for the FMP is compared with the
minimum value for FM, and causes it to fail.

The fix corrects the bug.

(cherry picked from commit 25564f73dcdfdd1de997d3f994c757d773623f0c)

Original-Signed-off-by: Hu Dou <hugh.dou@gmail.com>
GitOrigin-RevId: 25564f73dcdfdd1de997d3f994c757d773623f0c
Cr-Build-Id: 8716982175448322353
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8716982175448322353
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I1a5cb1ce5b2266fd5dc2ba608b0e48014da8c735
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6474697
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Keith Short <keithshort@chromium.org>