soc: silabs: siwx91x: Rename pins
The reference manual and the schematics of the reference designs
numbers the pins from 0 to 57. So let's also apply this scheme to the
pinctrl.

Note while PIN are numbered between 0 and 64, the GPIOs are represented
with 4 block of 16 GPIOs. This is the way the GPIOs are documented in
the reference manual, so we don't change that (In addition,
gpio_port_pins_t is a uint32_t, so it won't be possible to represent 64
GPIOs in one gpio_port_pins_t anyway).

(cherry picked from commit a1f48216bc7d1da86d527a4c681e069906042e71)

Original-Signed-off-by: Jérôme Pouiller <jerome.pouiller@silabs.com>
GitOrigin-RevId: a1f48216bc7d1da86d527a4c681e069906042e71
Cr-Build-Id: 8714785199454382401
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8714785199454382401
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: I30e00963bfc94d4d7d200463a6d3d03d2fb4b205
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6553667
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Copybot <copybot@chops-service-accounts.iam.gserviceaccount.com>