samples/drivers/mbox: Fix test regex
In 9d7ee123a54a920455322fa42a094c53e45e06d4
the regex used to verify the sample is doing what it should was changed
in a way in which it cannot possibly pass.
Let's fix it.

While at it, in the same commit, the simulator version of the test was
merged with the real_hw version, but the test name kept "real_hw" which
is now missleading => let's fix it.

And let's set as integration platform the simulated target, so it
can actually be runtime tested in CI.

(cherry picked from commit 72502938b34089e0abe40b8cd3b07dbca6f9342a)

Original-Signed-off-by: Alberto Escolar Piedras <alberto.escolar.piedras@nordicsemi.no>
GitOrigin-RevId: 72502938b34089e0abe40b8cd3b07dbca6f9342a
Cr-Build-Id: 8719722733849575601
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8719722733849575601
Copybot-Job-Name: zephyr-main-copybot-downstream
Change-Id: Iaecc3a7506b65d6f840d0658e4d7119e76f1339e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/6384751
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>